% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfscaling.R
\name{dfscaling}
\alias{dfscaling}
\title{Apply standard scaling and centering to the numeric features of a given input dataframe.}
\usage{
dfscaling(df)
}
\arguments{
\item{df}{Data Frame}
}
\value{
scaled_df Data Frame
}
\description{
The standard score of a sample x is calculated as:
z = (x - u) / s
where u is the mean of the training samples, and s is the standard deviation of the training samples.
}
\examples{
dfscaling(df)
}
