% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dftype.R
\name{dftype}
\alias{dftype}
\title{Explore the type of data frame variables and columns.}
\usage{
dftype(df)
}
\arguments{
\item{df}{Data Frame}
}
\value{
summary Vector

unique_df Data Frame
}
\description{
Create a new data frame which contains the type of columns and variables for the input data frame.
Furthermore, if there are non-numeric columns, it will return the unique values of non-numeric columns and their length.
}
\examples{
df <- data.frame(type = (c("Air", "Ship", "Bus", "Air")),
time= c(6, 32, 31, 5),
origin= c("US", "Mexico", "CANADA", "UK"))

results <- dftype(df)
results$summary
results$unique_df

}
