% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_data.R
\name{supervised_data}
\alias{supervised_data}
\title{Data Splitting for Supervised Machine Learning}
\usage{
supervised_data(data, xcols, ycols, ...)
}
\arguments{
\item{data}{the original dataset to be used for splitting}

\item{xcols}{a vector containing feature names (X) to be used as independent variables}

\item{ycols}{a vector containing target names (y) to be used as dependent variables or labels}

\item{...}{Additional parameters to pass to \code{tidymodels}'s \code{initial_split} function. See \code{tidymodels} \link[=https://www.rdocumentation.org/packages/rsample/versions/0.0.9/topics/initial_split]{documentation} for more details}
}
\value{
A list of the following components.
\itemize{
\item data - The original dataset unchanged
\item train - The training portion of the dataset
\item test - The test portion of the dataset
\item xtrain - The training portion of the dataset containing \code{X} features only.
\item ytrain - The training portion of the dataset containing \code{y} targets only.
\item xtest - The test portion of the dataset containing \code{X} features only.
\item ytest - The test portion of the dataset containing \code{y} targets only.
}
}
\description{
A function that utilizes \code{tidymodels}'s \code{initial_split} function to perform data spltting
while providing convenient access to \code{X} and \code{y} portions of both the test split and the train split.
}
\examples{
set.seed(1353)
cars <- supervised_data(mtcars, xcols = ['mpg', 'cyl', 'disp'], ycols=['hp'])
train_data <- cars$train
test_data <- cars$test
x_train <- cars$xtrain
y_train <- cars$ytrain
x_test <- cars$xtest
y_test <- cars$ytest
}
