% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoimpute_na.R
\name{autoimpute_na}
\alias{autoimpute_na}
\title{Identify and impute missing values in a dataframe.}
\usage{
autoimpute_na(df)
}
\arguments{
\item{df}{a dataframe or a tibble}
}
\value{
a dataframe or a tibble
}
\description{
Identify and impute missing values in a dataframe and return a new dataframe with imputed missing values.
}
\examples{
toy_df_na <- data.frame(chocolate_brand = (c("Lindt", "Rakhat", "Lindt",
                                             "Richart", "not available")),
                        price = c(3, NA, 4, 6, 3))
autoimpute_na(toy_df_na)
}
