% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfscaling.R
\name{dfscaling}
\alias{dfscaling}
\title{Apply standard scaling and centering to the numeric features of a given
input dataframe.}
\usage{
dfscaling(df, target)
}
\arguments{
\item{df}{Data Frame}

\item{target}{name of the target column}
}
\value{
scaled_df Data Frame
}
\description{
Each of the numeric columns will have a mean of 0
and standard deviation of 1 after the transformation.
}
\examples{
df <- tidyr::tibble(x = 1:5,
                      y = seq(2,10, by = 2),
                      z = x ^ 2 + y,
                      target = c(1,1,0,1,0))

dfscaling(df, target)
}
